+++ src/Views/TelaTED.java
@@ -182,7 +182,7 @@ public class TelaTED extends javax.swing.JFrame {
         //
         //
         if (tela.confirma()){
-            int saldo =usuario.getConta().getSaldo();
+            double saldo =usuario.getConta().getSaldo();
             int valor=Integer.parseInt(campoValor.getText());
             if (saldo >= valor){
                 if (valor>1000 && valor < 30000){
